Fix package API contracts and pricing
