+++ src/features/api/packages.ts
@@ -94,8 +94,10 @@ const loadPackageContext = async (
  * guard must reject a limited IP without building a package tree). */
 const loadPackageContextOr404 = async (
   slug: string,
-): Promise<PackageContext | Response> =>
-  (await loadPackageContext(slug)) ?? apiError(PACKAGE_NOT_FOUND, 404);
+): Promise<PackageContext | Response> => {
+  const loaded = await loadPackageContext(slug);
+  return loaded === null ? apiError(PACKAGE_NOT_FOUND, 404) : loaded;
+};
 
 /** Load a bookable package by slug, or respond with the package-not-found 404 —
  * shared by the GET and POST package endpoints via {@link withSlugLoaded} so the
@@ -103,6 +105,11 @@ const loadPackageContextOr404 = async (
 const withPackageContext = withSlugLoaded<PackageContext>(
   loadPackageContextOr404,
 );
+
+const packageChildren = (ctx: TicketCtx, memberId: number) => {
+  const children = ctx.childrenByParentId.get(memberId);
+  return children === undefined ? [] : children;
+};
 
 /** The contact-field requirement a package booking can validate against: the
  * members' settings merged with their children's (a chosen add-on can add a
@@ -112,9 +119,7 @@ const packageMergedFields = (ctx: TicketCtx): string =>
   mergeListingFields(
     ctx.listings.flatMap((e) => [
       e.listing.fields,
-      ...(ctx.childrenByParentId.get(e.listing.id) ?? []).map(
-        (c) => c.listing.fields,
-      ),
+      ...packageChildren(ctx, e.listing.id).map((c) => c.listing.fields),
     ]),
   );
 
@@ -140,7 +145,7 @@ export const handleGetPackage = withPackageContext(
     const members = group.hide_package_listings
       ? undefined
       : ctx.listings.map((e) => {
-          const children = (ctx.childrenByParentId.get(e.listing.id) ?? [])
+          const children = packageChildren(ctx, e.listing.id)
             .filter((child) => child.listing.active)
             .map((child) =>
               resolvedToPublicListing(
@@ -164,7 +169,7 @@ export const handleGetPackage = withPackageContext(
         maxPurchasable: limit,
         name: group.name,
         slug: group.slug,
-        ...(ctx.dates.length > 0 ? { availableDates: ctx.dates } : {}),
+        ...(ctx.dates.length ? { availableDates: ctx.dates } : {}),
         ...(customisable
           ? {
               dayCounts: dayCounts.map((days) => ({
@@ -172,7 +177,9 @@ export const handleGetPackage = withPackageContext(
                 priceMinor: packageBundleTotal(tree, days, bookableChildren),
               })),
             }
-          : { priceMinor: packageBundleTotal(tree, 1, bookableChildren) }),
+          : {
+              priceMinor: packageBundleTotal(tree, undefined, bookableChildren),
+            }),
         ...(members ? { members } : {}),
       },
     });
@@ -231,7 +238,9 @@ const resolvePackageOrder = async (
     ]),
   );
 
-  const dateResult = resolvePageDate(ctx.dates, String(body.date ?? ""));
+  const rawDate = body.date;
+  const submittedDate = rawDate === undefined ? null : String(rawDate);
+  const dateResult = resolvePageDate(ctx.dates, submittedDate);
   if (!dateResult.ok) return apiError(dateResult.error);
   const date = dateResult.date;
 
@@ -319,7 +328,7 @@ export const handleBookPackage = async (
     return finishFoldedBooking(
       request,
       form,
-      items.some((item) => item.unitPrice > 0),
+      items.some((item) => Boolean(item.unitPrice)),
       { date, fold, items },
     );
   });
